Add email signup, cleanup, duplicate at bottom
I thought it could be useful to promote email signup here, for users who want to be the first to hear about new features as we progress through the road map. I added a space between the icons and the text. I also duplicated the roadmap and email updates at the bottom of the page, in case the user scrolls to the bottom and hasn't found them. That's a little bit weird, but I'm slightly concerned that users will miss those links, or land on this page and think this page is the roadmap and not realize they need to go to a new page.